Add dh-python pinning to get it from upstream
The problem is that for some reason dh-python
package is put in MOS repo. When we create
partial ubuntu mirror we do not put dh-python
package into the partial mirror. But then
when we use this partial mirror together with
debootstrap (which does not support multiple mirrors)
the base operating system can not be installed
because python3 package requires dh-python.

Besides, those packages which are listed explicitly
in fuel_agent/drivers/nailgun.py#L602-L644 should
be also explicitly mentioned in requirements-deb.txt.

Change-Id: I370a37dbe056d426b4f043893e279e4c869bdf11
Closes-Bug: #1484434