Check and monitoring tool for repo consistency
In current workflow for every MOS project in review.fuel-infra.org we
first merge the patch to the source repository and then trigger the
job which publishes packages to the packages repository (deb- and rpm-
repos).
There are many circumstances which can cause publishing failures:
network issues, hardware problems, software misconfiguration, etc. So
this tool verify the content of the package repository is aligned with
the code merged in Gerrit.

Change-Id: I1ed0768b2e9261260c3b164e117cd057ae3dc817