Option to override version length in convert_version.py
Fuel projects use release segment consisting of 2 components rather
than 3. So for example 10.0.0b1 will be released as 10.0. This patch
adds option to convert_version.py to define number of components in
release segment. It also adds a list of fuel projects and overrides
version list for them during build.

Change-Id: I0b519e620ee0e7f3173b126e9e6c9766d87cea4e